re-activated skipped tests

+++ test/integration/api-key-integration.js
@@ -2,7 +2,6 @@ import { expect } from 'chai'
 import { after, before, describe, test } from 'mocha'
 import { client, createTestSpace, generateRandomId } from '../helpers'
 
-// check
 describe('ApiKey api', () => {
   let space
 
@@ -11,7 +10,9 @@ describe('ApiKey api', () => {
   })
 
   after(async () => {
-    return space.delete()
+    if (space) {
+      return space.delete()
+    }
   })
 
   test('Gets apiKeys', async () => {
