+++ test/integration/role-integration.js
@@ -32,7 +32,9 @@ describe('Role Api', function () {
   })
 
   after(async () => {
-    return space.delete()
+    if (space) {
+      return space.delete()
+    }
   })
 
   test('Gets roles', async () => {
